refactor(096-W4): BlackboardPage 拆分（1/4）——设置弹窗/Wiki 数据族/防抖订阅三线下沉（1286→855 行）
- blackboard/ 子目录新建：types.ts（3 接口）+ api.ts（3 fetch 函数，顺带统一
  WikiViewPage 的逐字重复定义）+ BlackboardSettingsModal.tsx（设置族 8 state +
  保存/恢复逻辑 + 两 Tab 子组件整组下沉，主组件仅剩 open 开关）
- hooks/useBlackboardWiki.ts：Wiki 数据族 5 state + latest-wins 竞态 ref 族 +
  fetch/清空/URL 同步 useEffect 族整段下沉（函数体逐字搬迁）
- hooks/useBlackboardDebounceStatus.ts：MobileDebounceIndicator ≡ BlackboardDebounceBar
  两处逐字订阅（9 行×2）收敛为单份 hook + 4 项单测
- 删除 useStateBlackboardData 无价值包装（useState 直接转发，YAGNI）
- 验证：tsc 零错误、vitest 366 全过（+4 新增）、Playwright 冒烟 4/4 过
  （页面装配/列表加载/弹窗开合/保存链路，dev 实例实测）
